Refactor : 컨트롤러 테스트 리펙토링
컨트롤러 테스트 util 을 mock 용도와 integrated test 용도로 나누고 리펙토링하였습니다.
