+++ src/test/java/nbbang/com/nbbang/global/support/controller/ControllerUtilMessage.java
@@ -1,0 +1,9 @@
+package nbbang.com.nbbang.global.support.controller;
+
+public abstract class ControllerUtilMessage {
+    public static String EXPECT_OK = "Expected OK, but wasn't";
+    public static String EXPECT_ERROR = "Expected Error, but wasn't";
+
+    public static String ILLEGAL_RESPONSE = "Failed to get Data";
+    public static String ILLEGAL_DEFAULT_RESPONSE = "Failed to get Data from Default Response";
+}
